reorganises code to follow structure in other examples
